requirements.txt
flask==2.0.1
python-ldap==3.3.0
jsonschema==3.2.0